// src/type.ts
export type Filters = {
    title: string;           // 과목명
    professor: string;       // 교수명
    courseNo: string;        // 과목 코드
    classroom: string;       // 강의실
    major: string;           // 전공
    grade: string;           // 학년
    times: string;           // 시간 필터
    courseTypeFilter: string; // 구분 필터
    credit: string;          // 학점
  };
  